aulavirtuallixfux/requirements.txt
asgiref==3.5.2
boto3==1.24.0
botocore==1.27.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.12.0
django-environ==0.8.1
django-heroku==0.3.1
django-oauth-toolkit==2.0.0
django-on-heroku==1.1.2
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
drf-social-oauth2==1.2.1
ecdsa==0.17.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
jwcrypto==1.3.1
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0
wrapt==1.14.1
